requirements.txt
aniso8601==8.0.0
astroid==2.3.3
autopep8==1.4.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.18.1
pandas==0.25.3
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
